feat: AccountRecordクラスをdataモジュールに移動させました
