+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/DataBase.kt
@@ -34,7 +34,8 @@ import net.pantasystem.milktea.data.infrastructure.url.db.UrlPreviewDAO
 import net.pantasystem.milktea.data.infrastructure.user.UserNicknameDAO
 import net.pantasystem.milktea.data.infrastructure.user.UserNicknameDTO
 import net.pantasystem.milktea.data.infrastructure.user.db.*
-import net.pantasystem.milktea.model.account.AccountInstanceTypeConverter
+import net.pantasystem.milktea.data.infrastructure.account.db.AccountInstanceTypeConverter
+import net.pantasystem.milktea.data.infrastructure.account.db.AccountRecord
 import net.pantasystem.milktea.model.notes.reaction.history.ReactionHistory
 import net.pantasystem.milktea.model.notes.reaction.history.ReactionHistoryDao
 import net.pantasystem.milktea.model.notes.reaction.usercustom.ReactionUserSetting
@@ -54,7 +55,7 @@ import net.pantasystem.milktea.model.url.UrlPreview
         DraftNoteDTO::class,
 
         UrlPreview::class,
-        net.pantasystem.milktea.model.account.AccountRecord::class,
+        AccountRecord::class,
         net.pantasystem.milktea.model.account.page.Page::class,
         MetaDTO::class,
         EmojiDTO::class,
