+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/drive/DriveFileRecord.kt
@@ -2,7 +2,7 @@ package net.pantasystem.milktea.data.infrastructure.drive
 
 import androidx.room.*
 import kotlinx.datetime.Instant
-import net.pantasystem.milktea.model.account.AccountRecord
+import net.pantasystem.milktea.data.infrastructure.account.db.AccountRecord
 import net.pantasystem.milktea.model.drive.FileProperty
 import net.pantasystem.milktea.model.user.User
